Move error-related stuff into its own file and merge with color
