+++ src/mars.h
@@ -346,27 +346,7 @@ enum MATCH
 
 typedef uint64_t StorageClass;
 
-
-void warning(Loc loc, const char *format, ...);
-void deprecation(Loc loc, const char *format, ...);
-void error(Loc loc, const char *format, ...);
-void errorSupplemental(Loc loc, const char *format, ...);
-void verror(Loc loc, const char *format, va_list ap, const char *p1 = NULL, const char *p2 = NULL, const char *header = "Error: ");
-void vwarning(Loc loc, const char *format, va_list);
-void verrorSupplemental(Loc loc, const char *format, va_list ap);
-void vdeprecation(Loc loc, const char *format, va_list ap, const char *p1 = NULL, const char *p2 = NULL);
-
-#if defined(__GNUC__) || defined(__clang__)
-__attribute__((noreturn))
-void fatal();
-#elif _MSC_VER
-__declspec(noreturn)
-void fatal();
-#else
-void fatal();
-#endif
-
-void halt();
+#include "errors.h"
 
 class Dsymbol;
 class Library;
